refactor subquery history implementation
