+++ fearless/Common/Network/Wallet/WalletRemoteHistoryProtocols.swift
